requirements.txt
flask == 3.1.1
python-dotenv == 1.1.1
pymongo == 4.13.2
paho-mqtt == 2.1.0
flasgger == 0.9.7.1
pydantic == 2.11.7
flask-cors == 6.0.1
pytest == 8.4.1
pytest-cov == 6.2.1
pdoc == 15.0.4